Feature/82 카카오 로그인 기능을 구현한다.

+++ src/main/java/com/grasshouse/dorandoran/comment/service/CommentLikeService.java
@@ -1,5 +1,7 @@
 package com.grasshouse.dorandoran.comment.service;
 
+import static com.grasshouse.dorandoran.common.exception.MemberMismatchException.COMMENT_LIKER_MISMATCH_MESSAGE;
+
 import com.grasshouse.dorandoran.comment.domain.Comment;
 import com.grasshouse.dorandoran.comment.domain.CommentLike;
 import com.grasshouse.dorandoran.comment.repository.CommentLikeRepository;
@@ -8,6 +10,8 @@ import com.grasshouse.dorandoran.comment.service.dto.CommentLikeCreateRequest;
 import com.grasshouse.dorandoran.common.exception.CommentLikeAlreadyExistsException;
 import com.grasshouse.dorandoran.common.exception.CommentLikeNotFoundException;
 import com.grasshouse.dorandoran.common.exception.CommentNotFoundException;
+import com.grasshouse.dorandoran.common.exception.MemberMismatchException;
+import com.grasshouse.dorandoran.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,14 +24,14 @@ public class CommentLikeService {
     private final CommentRepository commentRepository;
 
     @Transactional
-    public Long createCommentLike(CommentLikeCreateRequest request) {
+    public Long createCommentLike(CommentLikeCreateRequest request, Member member) {
         Comment comment = commentRepository
             .findById(request.getCommentId())
             .orElseThrow(CommentNotFoundException::new);
-        validateCommentLikeDuplication(request.getMemberId(), comment);
+        validateCommentLikeDuplication(member.getId(), comment);
 
         CommentLike commentLike = CommentLike.builder()
-            .memberId(request.getMemberId())
+            .memberId(member.getId())
             .comment(comment)
             .build();
 
@@ -42,9 +46,12 @@ public class CommentLikeService {
     }
 
     @Transactional
-    public void deleteCommentLike(Long commentLikeId) {
+    public void deleteCommentLike(Long commentLikeId, Member member) {
         CommentLike commentLike = commentLikeRepository.findById(commentLikeId)
             .orElseThrow(CommentLikeNotFoundException::new);
+        if (!commentLike.isSameLiker(member)) {
+            throw new MemberMismatchException(COMMENT_LIKER_MISMATCH_MESSAGE);
+        }
         commentLikeRepository.delete(commentLike);
     }
 }
